Fix bug in fb2-parser.ts
This commit fixes a bug in the fb2-parser.ts file where the cleanNonRussianWords function was not being called on all author names. The bug was caused by an incorrect regular expression that matched non-Russian characters. To fix this, I have modified the regular expression to match only Russian characters and added a new cleanNonRussianWords function that cleans up non-Russian words in author names.